start getting resource shared_ptr out of passes

+++ BasicRenderer/include/Managers/Singletons/DeletionManager.h
@@ -32,15 +32,15 @@ public:
 
 	void ProcessDeletions() {
 		m_deletionQueue.back().clear();
-		for (size_t i = m_deletionQueue.size() - 1; i >= 1; --i) {
+		for (int i = static_cast<int>(m_deletionQueue.size()) - 1; i >= 1; --i) {
 			m_deletionQueue[i].swap(m_deletionQueue[i-1]);
 		}
 		m_allocationDeletionQueue.back().clear();
-		for (size_t i = m_allocationDeletionQueue.size() - 1; i >= 1; --i) {
+		for (int i = static_cast<int>(m_allocationDeletionQueue.size()) - 1; i >= 1; --i) {
 			m_allocationDeletionQueue[i].swap(m_allocationDeletionQueue[i-1]);
 		}
 		m_trackedAllocationDeletionQueue.back().clear();
-		for (size_t i = m_trackedAllocationDeletionQueue.size() - 1; i >= 1; --i) {
+		for (int i = static_cast<int>(m_trackedAllocationDeletionQueue.size()) - 1; i >= 1; --i) {
 			m_trackedAllocationDeletionQueue[i].swap(m_trackedAllocationDeletionQueue[i-1]);
 		}
 	}
